Migrated Stat Fill into Page Class
There's a lot here, and we want to move to a per page fill in
